all Controllers done

+++ app/Http/Controllers/RegistrationController.php
@@ -4,12 +4,19 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use App\Meeting;
+use App\User;
+
 class RegistrationController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('jwt.auth');
+    }
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -22,26 +29,30 @@ class RegistrationController extends Controller
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
 
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
+        $meeting = Meeting::findOrFail($meeting_id);
+        $user = User::findOrFail($user_id);
+
+        $message = [
+            'msg' => 'User is already registered from meeting',
+            'user' => $user,
+            'meeting' => $meeting,
+            'unregister' => [
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
+                'method' => 'DELETE',
             ]
         ];
+        if ($meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json($message, 404);
+        };
 
-        $user = [
-            'name' => 'Name'
-        ];
+        $user->meetings()->attach($meeting);
 
         $response = [
             'msg' => 'User registered for meeting',
             'meeting' => $meeting,
             'user' => $user,
             'unregister' => [
-                'href' => 'api/v1/meeting/registration/1',
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
                 'method' => 'DELETE'
             ]
         ];
@@ -52,24 +63,20 @@ class RegistrationController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
+        $meeting = Meeting::findOrFail($id);
+        if (! $user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
+        if (!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, delete operation not successful'], 401);
+        };
 
-        $user = [
-            'name' => 'Name'
-        ];
+        $meeting->users()->detach($user->id);
 
         $response = [
             'msg' => 'User unregistered for meeting',
